+++ rust/services/call/host/src/db.rs
@@ -1,7 +1,0 @@
-pub mod proof;
-pub mod provider;
-use call_common::RevmDB;
-use revm::db::CacheDB;
-
-pub type HostDb = CacheDB<proof::ProofDb>;
-pub type HostDbError = <HostDb as RevmDB>::Error;