+++ rust/guest_wrapper/artifacts/chain_guest/CHANGELOG.md
@@ -19,3 +19,4 @@
   * `3dedd7024c3462f3f326489b27ff1843f30f03afff71cdd88ccd24f81fbe8f67` – Deny panics globally [#2199](https://github.com/vlayer-xyz/vlayer/pull/2199)
   * `304957ee9fdecb19ec65f8633d40a926aacd142b81c4cfd43d42ca6742f67130` – Update `chain_specs.toml` [#2218](https://github.com/vlayer-xyz/vlayer/pull/2218)
   * `d47c721762613c905c892b615c3496fc172ea7d5f7fc29ae9900a4f0b5478dd3` – Risc0 2.0.2 - [#2279](https://github.com/vlayer-xyz/vlayer/pull/2279)
+  * `202097073e66ede38bf441f210b93e7983bd2473b1680f945ebacf41bf1d5ed8` – Update `AnchorStateRegistry` addresses for unichain and base chains